refactor: Member 위치 변경에 따른 import path 수정
